requirements.txt
python-telegram-bot[job-queue]==20.7
python-dotenv==1.0.0
APScheduler==3.10.4
pydantic==2.5.0
python-dateutil==2.8.2
cryptography==42.0.4
httpx==0.25.2